Flag to delete workdir

+++ src/python/merge_tester.py
@@ -10,6 +10,7 @@ SCRATCH_DIR = "scratch/"
 STORE_SCRATCH = False
 WORKDIR = ".workdir/"
 CACHE = "cache/merge_test_results/"
+DELETE_WORKDIR = True
 
 
 def get_repo(repo_name):
@@ -123,13 +124,15 @@ def test_merge(args):
                 +str(intelli_merge)+" "+str(int(intelli_runtime))
         with open(cache_file,"w") as f:
             f.write(out)
-        shutil.rmtree(repo_dir_copy)
+        if DELETE_WORKDIR:
+            shutil.rmtree(repo_dir_copy)
         return git_merge,git_runtime,spork_merge,spork_runtime,intelli_merge,intelli_runtime
     except Exception:
         with open(cache_file,"w") as f:
             out = "-1 -1 -1 -1 -1 -1"
             f.write(out)
-        shutil.rmtree(repo_dir_copy)
+        if DELETE_WORKDIR:
+            shutil.rmtree(repo_dir_copy)
         return -1,-1,-1,-1,-1,-1
